requirements.txt
streamlit==1.32.0
openai==1.12.0
python-dotenv==1.0.0
pandas==2.2.0 